+++ src/classes/com/sun/tdk/jcov/instrument/InstrumentationParams.java
@@ -26,6 +26,7 @@ package com.sun.tdk.jcov.instrument;
 
 import com.sun.tdk.jcov.instrument.InstrumentationOptions.ABSTRACTMODE;
 import com.sun.tdk.jcov.instrument.InstrumentationOptions.InstrumentationMode;
+import com.sun.tdk.jcov.instrument.plugin.FilteringPlugin;
 import com.sun.tdk.jcov.runtime.Collect;
 import com.sun.tdk.jcov.runtime.CollectDetect;
 import com.sun.tdk.jcov.util.Utils;
@@ -208,9 +209,7 @@ public class InstrumentationParams {
     }
 
     public InstrumentationPlugin filter(InstrumentationPlugin plugin) {
-        //TODO is it possible to optimize to return the original plugin
-        //in case for filtering options are provided?
-        return new InstrumentationPlugin.FilteringPlugin(plugin, this::isIncluded);
+        return new FilteringPlugin(plugin, this::isIncluded);
     }
 
     public boolean isDetectInternal() {
