+++ src/classes/com/sun/tdk/jcov/instrument/InstrumentationPlugin.java
@@ -24,36 +24,16 @@
  */
 package com.sun.tdk.jcov.instrument;
 
-import com.sun.tdk.jcov.util.Utils;
-
-import java.io.Closeable;
-import java.io.IOException;
-import java.io.InputStream;
 import java.io.OutputStream;
-import java.io.UncheckedIOException;
-import java.net.MalformedURLException;
-import java.net.URL;
-import java.net.URLClassLoader;
-import java.nio.file.FileSystem;
-import java.nio.file.FileSystems;
-import java.nio.file.Files;
-import java.nio.file.Path;
-import java.util.ArrayList;
 import java.util.Collection;
-import java.util.List;
 import java.util.Map;
-import java.util.ServiceLoader;
 import java.util.function.BiConsumer;
 import java.util.function.Consumer;
-import java.util.function.Predicate;
-import java.util.jar.JarFile;
-import java.util.stream.Collectors;
-import java.util.zip.ZipEntry;
-
-import static com.sun.tdk.jcov.util.Utils.isClassFile;
 
 /**
- * TODO describe the lifecycle
+ * A plugin's {@link #instrument(Collection, ClassLoader, BiConsumer, InstrumentationParams)} could be used
+ * multiple times to instrument additional bytecode. A plugin is supposed to accumulate the the, which later
+ * is supposed to be used through {@link #complete()} method.
  */
 public interface InstrumentationPlugin {
 
@@ -93,265 +73,5 @@ public interface InstrumentationPlugin {
 
     default boolean isClass(String resource) {
         return resource.endsWith(CLASS_EXTENTION) && !resource.endsWith(MODULE_INFO_CLASS);
-    }
-
-    //TODO properly relocate the inner classes
-
-    interface ModuleInstrumentationPlugin {
-        String getModuleName(byte[] moduleInfo);
-        byte[] addExports(List<String> exports, byte[] moduleInfo, ClassLoader loader);
-        byte[] clearHashes(byte[] moduleInfo, ClassLoader loader);
-    }
-
-    abstract class ProxyInstrumentationPlugin implements InstrumentationPlugin {
-        private final InstrumentationPlugin inner;
-
-        protected ProxyInstrumentationPlugin(InstrumentationPlugin inner) {
-            this.inner = inner;
-        }
-
-        public InstrumentationPlugin getInner() {
-            return inner;
-        }
-
-        @Override
-        public final Map<String, Consumer<OutputStream>> complete() throws Exception {
-            return inner.complete();
-        }
-    }
-
-    class FilteringPlugin extends ProxyInstrumentationPlugin {
-        private final Predicate<String> filter;
-
-        public FilteringPlugin(InstrumentationPlugin inner, Predicate<String> filter) {
-            super(inner);
-            this.filter = filter;
-        }
-
-        @Override
-        public void instrument(Collection<String> resources, ClassLoader loader,
-                               BiConsumer<String, byte[]> saver, InstrumentationParams parameters) throws Exception {
-            List<String> accepted = new ArrayList<>();
-            List<String> rejected = new ArrayList<>();
-            resources.forEach(r -> {
-                if (filter.test(r)) accepted.add(r);
-                else rejected.add(r);
-            });
-            getInner().instrument(accepted, loader, saver, parameters);
-            rejected.forEach(c -> {
-                try (InputStream in = loader.getResourceAsStream(c)) {
-                    saver.accept(c, in.readAllBytes());
-                } catch (IOException e) {
-                    throw new UncheckedIOException(e);
-                }
-            });
-        }
-    }
-
-    static Predicate<String> classNameFilter(InstrumentationParams params) {
-        return r -> {
-            if (isClassFile(r))
-                return params.isIncluded(r.substring(0,
-                        r.length() - Utils.FILE_TYPE.CLASS.name().length() - 1));
-            else return true;
-        };
-    }
-
-    interface Source extends Closeable {
-        //paths relative to the result root
-        Collection<String> resources() throws Exception;
-        ClassLoader loader();
-    }
-
-    class ImplantingPlugin extends ProxyInstrumentationPlugin {
-        private final Source source;
-
-        //TODO similar to ModuleImplantingPlugin have different implants for different locations somehow?
-        public ImplantingPlugin(InstrumentationPlugin inner, Source source) {
-            super(inner);
-            this.source = source;
-        }
-
-        @Override
-        public void instrument(Collection<String> classes, ClassLoader loader,
-                               BiConsumer<String, byte[]> saver, InstrumentationParams parameters) throws Exception {
-            getInner().instrument(classes, loader, saver, parameters);
-            for(String r : source.resources())
-                try (InputStream in = source.loader().getResourceAsStream(r)) {
-                    saver.accept(r, in.readAllBytes());
-                }
-        }
-    }
-
-    //TODO better be private
-    class OverridingClassLoader extends URLClassLoader {
-
-        private static URL[] toURL(Path root) {
-            try {
-                return new URL[] {root.toUri().toURL()};
-            } catch (MalformedURLException e) {
-                //should not happen since getting teh URL legally
-                throw new RuntimeException(e);
-            }
-        }
-
-        private final ClassLoader backup;
-
-        public OverridingClassLoader(Path root, ClassLoader backup) {
-            this(toURL(root), backup);
-        }
-
-        public OverridingClassLoader(URL[] urls, ClassLoader backup) {
-            super(urls);
-            this.backup = backup;
-        }
-
-        @Override
-        public URL getResource(String name) {
-            //first try to find local resource, from teh current module
-            URL resource = findResource(name);
-            //for module-info it does not make sense to look in other classloaders
-            if(name.equals(MODULE_INFO_CLASS)) return resource;
-            //if none, try other modules
-            if (resource == null) resource = backup.getResource(name);
-            //that should not happen during normal use
-            //if happens, refer to super, nothing else we can do
-            if (resource == null) resource = super.getResource(name);
-            return resource;
-        }
-    }
-
-    interface Destination extends Closeable {
-        BiConsumer<String, byte[]> saver();
-    }
-
-    /**
-     * A utility class which works with a given plugin in turms of file hierarchies.
-     */
-    class Instrumentation {
-        private final InstrumentationPlugin inner;
-
-        public Instrumentation(InstrumentationPlugin inner) {
-            this.inner = inner;
-        }
-
-        public void instrument(Source source, Destination destination,
-                                     InstrumentationParams parameters) throws Exception {
-            inner.instrument(source.resources(), source.loader(),
-                    destination.saver(), parameters);
-        }
-    }
-
-    /**
-     * Helps to instrument modules.
-     * @see #Instrumentation
-     */
-    class ModuleInstrumentation extends Instrumentation {
-        private final ModuleInstrumentationPlugin modulePlugin;
-
-        public ModuleInstrumentation(InstrumentationPlugin inner, ModuleInstrumentationPlugin modulePlugin) {
-            super(inner);
-            this.modulePlugin = modulePlugin;
-        }
-
-        public ModuleInstrumentationPlugin getModulePluign() {
-            return modulePlugin;
-        }
-
-        /**
-         * Take any required action needed to instrument a module. This implementation does not do anything.
-         * @param moduleInfo
-         * @param loader
-         * @param destination
-         * @throws Exception
-         * @see ModuleInstrumentationPlugin
-         */
-        protected void proccessModule(byte[] moduleInfo, ClassLoader loader, BiConsumer<String, byte[]> destination)
-                throws Exception {
-        }
-
-        @Override
-        public void instrument(Source source, Destination destination,
-                               InstrumentationParams parameters) throws Exception {
-            super.instrument(source, destination, parameters);
-            try (InputStream in = source.loader().getResourceAsStream(MODULE_INFO_CLASS)) {
-                byte[] moduleInfo = in.readAllBytes();
-                proccessModule(moduleInfo, source.loader(), destination.saver());
-            }
-        }
-    }
-
-    class PathSource implements Source, Closeable {
-
-        private final ClassLoader loader;
-        private final Path root;
-
-        public PathSource(ClassLoader backup, Path root) {
-            this.loader = new OverridingClassLoader(root, backup);
-            this.root = root;
-        }
-
-        @Override
-        public Collection<String> resources() throws Exception {
-            if(Files.isDirectory(root))
-                return Files.find(root, Integer.MAX_VALUE, (f, a) -> Files.isRegularFile(f))
-                        .map(r -> root.relativize(r).toString())
-                        .collect(Collectors.toList());
-            else
-                try (JarFile jar = new JarFile(root.toFile())) {
-                    return jar.stream().filter(f -> !f.isDirectory())
-                            .map(ZipEntry::getName).collect(Collectors.toList());
-                }
-        }
-
-        @Override
-        public ClassLoader loader() {
-            return loader;
-        }
-
-        @Override
-        public void close() throws IOException {
-            if (loader instanceof Closeable) ((Closeable) loader).close();
-        }
-
-        public boolean isModule() throws IOException {
-            if (Files.isDirectory(root)) {
-                return Files.exists(root.resolve(MODULE_INFO_CLASS));
-            } else {
-                try (JarFile jar = new JarFile(root.toFile())) {
-                    return jar.stream().map(ZipEntry::getName).anyMatch(MODULE_INFO_CLASS::equals);
-                }
-            }
-        }
-    }
-
-    class PathDestination implements Destination, Closeable {
-        private final Path root;
-        private final FileSystem fs;
-        private final BiConsumer<String, byte[]> saver;
-
-        public PathDestination(Path root) throws IOException {
-            fs = Files.isDirectory(root) ? null : FileSystems.newFileSystem(root, (ClassLoader) null);
-            this.root = Files.isDirectory(root) ? root : fs.getPath("/");
-            saver = (s, bytes) -> {
-                try {
-                    Path f = PathDestination.this.root.resolve(s);
-                    Files.createDirectories(f.getParent());
-                    Files.write(f, bytes);
-                } catch (IOException e) {
-                    throw new UncheckedIOException(e);
-                }
-            };
-        }
-
-        @Override
-        public void close() throws IOException {
-            if (fs != null) fs.close();
-        }
-
-        @Override
-        public BiConsumer<String, byte[]> saver() {
-            return saver;
-        }
     }
 }
