+++ src/classes/com/sun/tdk/jcov/instrument/plugin/PathSource.java
@@ -1,0 +1,80 @@
+/*
+ * Copyright (c) 2023 Oracle and/or its affiliates. All rights reserved.
+ * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
+ *
+ * This code is free software; you can redistribute it and/or modify it
+ * under the terms of the GNU General Public License version 2 only, as
+ * published by the Free Software Foundation.  Oracle designates this
+ * particular file as subject to the "Classpath" exception as provided
+ * by Oracle in the LICENSE file that accompanied this code.
+ *
+ * This code is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or
+ * FITNESS FOR A PARTICULAR PURPOSE.  See the GNU General Public License
+ * version 2 for more details (a copy is included in the LICENSE file that
+ * accompanied this code).
+ *
+ * You should have received a copy of the GNU General Public License version
+ * 2 along with this work; if not, write to the Free Software Foundation,
+ * Inc., 51 Franklin St, Fifth Floor, Boston, MA 02110-1301 USA.
+ *
+ * Please contact Oracle, 500 Oracle Parkway, Redwood Shores, CA 94065 USA
+ * or visit www.oracle.com if you need additional information or have any
+ * questions.
+ */
+package com.sun.tdk.jcov.instrument.plugin;
+
+import com.sun.tdk.jcov.instrument.InstrumentationPlugin;
+
+import java.io.Closeable;
+import java.io.IOException;
+import java.nio.file.Files;
+import java.nio.file.Path;
+import java.util.Collection;
+import java.util.jar.JarFile;
+import java.util.stream.Collectors;
+import java.util.zip.ZipEntry;
+
+public class PathSource implements Source, Closeable {
+
+    private final ClassLoader loader;
+    private final Path root;
+
+    public PathSource(ClassLoader backup, Path root) {
+        this.loader = new OverridingClassLoader(root, backup);
+        this.root = root;
+    }
+
+    @Override
+    public Collection<String> resources() throws Exception {
+        if (Files.isDirectory(root))
+            return Files.find(root, Integer.MAX_VALUE, (f, a) -> Files.isRegularFile(f))
+                    .map(r -> root.relativize(r).toString())
+                    .collect(Collectors.toList());
+        else
+            try (JarFile jar = new JarFile(root.toFile())) {
+                return jar.stream().filter(f -> !f.isDirectory())
+                        .map(ZipEntry::getName).collect(Collectors.toList());
+            }
+    }
+
+    @Override
+    public ClassLoader loader() {
+        return loader;
+    }
+
+    @Override
+    public void close() throws IOException {
+        if (loader instanceof Closeable) ((Closeable) loader).close();
+    }
+
+    public boolean isModule() throws IOException {
+        if (Files.isDirectory(root)) {
+            return Files.exists(root.resolve(InstrumentationPlugin.MODULE_INFO_CLASS));
+        } else {
+            try (JarFile jar = new JarFile(root.toFile())) {
+                return jar.stream().map(ZipEntry::getName).anyMatch(InstrumentationPlugin.MODULE_INFO_CLASS::equals);
+            }
+        }
+    }
+}
